+++ pyrope/nodes/dtype_nodes.py
@@ -1,15 +1,13 @@
 
-import ast
 from fractions import Fraction
-import numpy as np
+
 import sympy
-import tokenize
 
 from pyrope import config
 from pyrope.dtypes import (
      BoolType, ComplexType, DictType, EquationType, ExpressionType, IntType,
-     ListType, MatrixType, OneOfType, RationalType, RealType, SetType,
-     StringType, TupleType, VectorType
+     LinearExpressionType, ListType, MatrixType, OneOfType, PolynomialType,
+     RationalType, RealType, SetType, StringType, TupleType, VectorType
 )
 from pyrope.errors import ValidationError
 from pyrope.nodes import Checkbox, Dropdown, Node, RadioButtons, Slider, Text
@@ -45,55 +43,11 @@ class Problem(Node):
         ) / len(self.ifields)
 
 
-class Parser(Node):
-
-    # c.f. https://docs.python.org/3/library/ast.html#ast.literal_eval
-    dtypes = (str, int, float, complex, tuple, list, dict, set, bool)
-
-    def __init__(self, ifield, dtype, **kwargs):
-        Node.__init__(self, '<<_>>', {'_': ifield}, **kwargs)
-        if dtype.dtype not in self.dtypes:
-            raise TypeError(f'Unable to parse data type {dtype.dtype}.')
-        self.dtype = dtype
-
-    # c.f. https://bugs.python.org/issue39159
-    @staticmethod
-    def safe_eval(s):
-        if len(s) > config.maximum_input_length:
-            raise ValidationError(
-                f'Input size {len(s)} '
-                f'exceeds limit {config.maximum_input_length}.'
-            )
-        try:
-            expr = ast.literal_eval(s)
-        except (MemoryError, SyntaxError, TypeError, ValueError) as e:
-            raise ValidationError(e)
-        return expr
-
-    def assemble(self, _):
-        if not isinstance(_, str) or self.dtype.dtype == str:
-            return _
-        if _ == '':
-            return None
-        value = self.safe_eval(_)
-        if value is None:
-            return _
-        return value
-
-    def disassemble(self, value):
-        return {'_': str(value)}
-
-
 class Bool(Node):
 
     def __init__(self, *, widget=Checkbox(), **kwargs):
         self.dtype = BoolType(**kwargs)
         Node.__init__(self, '<<_>>', {'_': widget}, **kwargs)
-
-    def assemble(self, _):
-        if _ == '':
-            return None
-        return _
 
 
 class Complex(Node):
@@ -107,7 +61,7 @@ class Complex(Node):
                 i_on_the=i_on_the, widget=widget, **kwargs
             )
         else:
-            _ = Parser(widget, self.dtype, **kwargs)
+            _ = widget
         Node.__init__(self, '<<_>>', {'_': _}, **kwargs)
 
 
@@ -139,55 +93,21 @@ class Dict(Node):
 
     def __init__(self, *, widget=Text(), **kwargs):
         self.dtype = DictType(**kwargs)
-        Node.__init__(
-            self, '<<_>>', {'_': Parser(widget, self.dtype, **kwargs)},
-            **kwargs
-        )
+        Node.__init__(self, '<<_>>', {'_': widget}, **kwargs)
 
 
 class Expression(Node):
 
-    def __init__(self, *, widget=Text(), transformations=None, **kwargs):
+    def __init__(self, *, widget=Text(), **kwargs):
         self.dtype = ExpressionType(**kwargs)
         Node.__init__(self, '<<_>>', {'_': widget}, **kwargs)
-        if transformations is None:
-            transformations = tuple(
-                getattr(sympy.parsing.sympy_parser, transformation)
-                for transformation in config.transformations
-            )
-        self.transformations = transformations
-
-    def assemble(self, _):
-        if _ == '':
-            return None
-        try:
-            expr = sympy.parse_expr(_, transformations=self.transformations)
-        except (SyntaxError, TypeError, tokenize.TokenError) as e:
-            raise ValidationError(e)
-        return expr
-
-    def disassemble(self, value):
-        return {'_': sympy.srepr(value)}
 
 
-class Equation(Expression):
+class Equation(Node):
 
     def __init__(self, *, widget=Text(), **kwargs):
-        Expression.__init__(self, widget=widget, **kwargs)
         self.dtype = EquationType(**kwargs)
-
-    def assemble(self, _):
-        if _ == '':
-            return None
-        if '=' not in _:
-            raise ValidationError('An equation needs an equal sign.')
-        if _.count('=') != 1:
-            raise ValidationError('Equation contains multiple equal signs.')
-        LHS, RHS = _.split('=')
-        return Expression.assemble(self, f'Eq({LHS}, {RHS}, evaluate=False)')
-
-    def disassemble(self, value):
-        return {'_': f'{sympy.srepr(value.lhs)}={sympy.srepr(value.rhs)}'}
+        Node.__init__(self, '<<_>>', {'_': widget}, **kwargs)
 
 
 class Int(Node):
@@ -202,10 +122,7 @@ class Int(Node):
                 widget = Slider(self.dtype.minimum, self.dtype.maximum)
             else:
                 widget = Text()
-        Node.__init__(
-            self, '<<_>>', {'_': Parser(widget, self.dtype, **kwargs)},
-            **kwargs
-        )
+        Node.__init__(self, '<<_>>', {'_': widget}, **kwargs)
 
 
 class Matrix(Node):
@@ -213,19 +130,6 @@ class Matrix(Node):
     def __init__(self, *, widget=Text(), **kwargs):
         self.dtype = MatrixType(**kwargs)
         Node.__init__(self, '<<_>>', {'_': widget}, **kwargs)
-
-    def assemble(self, _):
-        if _ == '':
-            return None
-        value = Parser.safe_eval(_)
-        try:
-            value = np.array(value)
-        except ValueError as e:
-            raise ValidationError(e)
-        return value
-
-    def disassemble(self, value):
-        return {'_': str(value.tolist())}
 
 
 class Vector(Matrix):
@@ -255,10 +159,89 @@ class OneOf(Node):
                 widget = RadioButtons(*args)
             else:
                 widget = Dropdown(*args)
+        Node.__init__(self, '<<_>>', {'_': widget}, **kwargs)
+
+
+class Polynomial(Node):
+
+    def __init__(
+            self, *, degree=None, elementwise=False, widget=Text(), **kwargs
+    ):
+        self.dtype = PolynomialType(degree=degree, **kwargs)
+        if elementwise is True:
+            _ = ElementwisePolynomial(degree, widget=widget, **kwargs)
+            Node.__init__(self, '<<_>>', {'_': _}, **kwargs)
+        else:
+            Node.__init__(self, '<<_>>', {'_': widget}, **kwargs)
+
+
+class ElementwisePolynomial(Node):
+
+    def __init__(self, degree, *, widget=Text(), **kwargs):
+        self.dtype = PolynomialType(degree=degree, elementwise=True, **kwargs)
+        self.degree = degree
+        if self.degree is None:
+            raise ValueError(
+                'Cannot render polynomials with an arbitrary degree '
+                'elementwise.'
+            )
+        if len(self.dtype.symbols) >= 2:
+            raise NotImplementedError(
+                'Cannot render polynomials with more than one symbol '
+                'elementwise.'
+            )
+        self.symbol = next(iter(self.dtype.symbols))
+        tokens = [
+            f'<<x_{i}>>${self.symbol}^{"{"}{i}{"}"}$'
+            for i in range(self.degree, 1, -1)
+        ]
+        if self.degree >= 1:
+            tokens.append(f'<<x_1>>${self.symbol}$')
+        tokens.append('<<x_0>>')
+        template = ' + '.join(tokens)
         Node.__init__(
-            self, '<<_>>', {'_': Parser(widget, self.dtype, **kwargs)},
-            **kwargs
+            self, template, {
+                f'x_{i}': Rational(elementwise=False, widget=widget, **kwargs)
+                for i in range(0, self.degree + 1)
+            }
         )
+
+    def assemble(self, **ifields):
+        return sympy.Poly.from_dict({
+            i: ifields[f'x_{i}'] for i in range(0, self.degree + 1)
+        }, self.symbol)
+
+    def disassemble(self, value):
+        coeffs = value.all_coeffs()[::-1]
+        return {
+            f'x_{i}': coeffs[i] for i in range(0, self.degree + 1)
+        }
+
+
+class LinearExpression(Node):
+
+    def __init__(self, *, elementwise=False, widget=Text(), **kwargs):
+        self.dtype = LinearExpressionType(elementwise=elementwise, **kwargs)
+        if elementwise is True:
+            _ = ElementwiseLinearExpression(widget=widget, **kwargs)
+            Node.__init__(self, '<<_>>', {'_': _}, **kwargs)
+        else:
+            Node.__init__(self, '<<_>>', {'_': widget}, **kwargs)
+
+
+class ElementwiseLinearExpression(ElementwisePolynomial):
+
+    def __init__(self, **kwargs):
+        ElementwisePolynomial.__init__(self, 1, **kwargs)
+        self.dtype = LinearExpressionType(elementwise=True, **kwargs)
+
+    def disassemble(self, value):
+        coeffs = value.all_coeffs()[::-1]
+        if value.degree() == 0:
+            coeffs.append(0)
+        return {
+            f'x_{i}': coeffs[i] for i in range(0, self.degree + 1)
+        }
 
 
 class Rational(Node):
@@ -272,20 +255,6 @@ class Rational(Node):
             _ = widget
         Node.__init__(self, '<<_>>', {'_': _}, **kwargs)
 
-    def assemble(self, _):
-        if _ == '':
-            return None
-        try:
-            fraction = Fraction(_)
-        except (ValueError, TypeError, ZeroDivisionError) as e:
-            raise ValidationError(e)
-        return fraction
-
-    def disassemble(self, value):
-        if self.elementwise:
-            return Node.disassemble(self, value)
-        return {'_': str(value)}
-
 
 class ElementwiseRational(Node):
 
@@ -294,7 +263,7 @@ class ElementwiseRational(Node):
         Node.__init__(
             self, '<<a>> / <<b>>', {
                 'a': Int(widget=widget, **kwargs),
-                'b': Int(widget=widget, **kwargs)
+                'b': Natural(widget=widget, with_zero=False, **kwargs)
             }, **kwargs
         )
 
@@ -309,20 +278,14 @@ class Real(Node):
 
     def __init__(self, *, widget=Text(), **kwargs):
         self.dtype = RealType(**kwargs)
-        Node.__init__(
-            self, '<<_>>', {'_': Parser(widget, dtype=self.dtype, **kwargs)},
-            **kwargs
-        )
+        Node.__init__(self, '<<_>>', {'_': widget}, **kwargs)
 
 
 class Set(Node):
 
     def __init__(self, *, widget=Text(), **kwargs):
         self.dtype = SetType(**kwargs)
-        Node.__init__(
-            self, '<<_>>', {'_': Parser(widget, self.dtype, **kwargs)},
-            **kwargs
-        )
+        Node.__init__(self, '<<_>>', {'_': widget}, **kwargs)
 
 
 class String(Node):
@@ -338,10 +301,7 @@ class Tuple(Node):
 
     def __init__(self, widget=Text(), **kwargs):
         self.dtype = self.dtype(**kwargs)
-        Node.__init__(
-            self, '<<_>>', {'_': Parser(widget, self.dtype, **kwargs)},
-            **kwargs
-        )
+        Node.__init__(self, '<<_>>', {'_': widget}, **kwargs)
 
 
 class List(Tuple):