+++ docs/changelog.rst
@@ -19,6 +19,7 @@ New
 * Weight scores with :code:`weights` when instantiating an exercise and test :code:`weights`.
 * Implement :code:`atol` and :code:`rtol` for :code:`Complex` and :code:`Real` nodes.
 * Add specific exercises from a Python script to an exercise pool via the CLI, i.e. :code:`/path/to/exercises/exercises.py:Example1,Example2,...`.
+* New nodes: :code:`Polynomial`, :code:`ElementwisePolynomial`, :code:`LinearExpression` and :code:`ElementwiseLinearExpression`.
 
 Changes
 -------
@@ -32,6 +33,7 @@ Changes
 * Validate arguments of widgets.
 * Rename :code:`score_types` to :code:`float_types`.
 * Create a :code:`nodes` package and outsource errors into a separate module to avoid circular imports.
+* Composed input fields can be invalid even if all children nodes are valid. Therefore all children nodes are now invalid if the composed input field is invalid.
 
 Fixes
 -----